requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
boto3==1.38.42
botocore==1.38.42
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cookiecutter==2.6.0
crispy-tailwind==1.0.3
diff-match-patch==20241021
Django==5.1.6
django-crispy-forms==2.4
django-import-export==4.3.7
django-tailwind==3.8.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openpyxl==3.1.5
packaging==24.2
pillow==11.3.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20250822
tzdata==2025.1
urllib3==2.5.0
whitenoise==6.9.0
